requirements.txt
boto3>=1.26.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0